The code introduces a centralised error handling mechanism for arithmetic operations in a custom assembly language interpreter. Here's a breakdown:
**Key Changes:**

1. **`call` Function in `Arithmetic.kt`:**
   - A new function `call` is added to the `Arithmetic` class.
   - It takes the operation name (`name`) and a lambda function (`function`) as parameters.
   - The lambda function encapsulates the actual arithmetic operation (e.g., addition, subtraction).
   - Inside `call`:
     - It executes the provided `function` within a `try-catch` block.
     - If an exception occurs, it throws a `GeneralArithmeticException` with the operation `name`.
     - It sets the sign flag (`SF`) and zero flag (`ZF`) based on the result in register `R4`.
     - It sets the error non-specific flag (`ENSF`) to indicate whether an error occurred (`true` for error, `false` otherwise).

2. **Refactoring Arithmetic Instructions:**
   - Functions like `add`, `sub`, `mul`, `div`, `mod`, `pow` are modified to use the `call` function.
   - They pass the operation name (e.g., "add") and a lambda containing the specific arithmetic logic to the `call` function.
   - Error handling and flag setting are now delegated to the `call` function.

3. **Error Specific Flag (`ESF`) in `IntelRegisters`:**
   - A new enum value `ESF` (Error Specific Flag) is added to the `IntelRegisters` enum.
   - This flag is intended to store specific error codes.

4. **`ErrorType` Enum:**
   - The `ErrorType` enum now uses `Long` for its `code` property instead of `Byte`.
   - A new error type `DIVIDE_BY_0` is added with code `3`.

**Example (Addition):**

```kotlin
fun Arithmetic.add(registerA: RegisterType, registerB: RegisterType) = call("add") {
    val a: Long = registers.read(register = registerA)
    val b: Long = registers.read(register = registerB)
    val out = a + b
    registers.write(RegisterType.R4, out)
}
```

**Benefits:**

- **Centralized Error Handling:**  Arithmetic errors are now handled consistently in one place (`call` function).
- **Cleaner Code:** Arithmetic instruction functions are more concise as error handling is delegated.
- **Specific Error Codes:** The `ESF` flag allows for storing specific error codes, enabling more informative error reporting.

**Potential Improvements:**

- **Detailed Error Messages:** The `GeneralArithmeticException` could include more context about the error (e.g., operand values).
- **Error Code Utilization:** The `ESF` flag is introduced but not yet fully utilized. The code could be extended to set specific error codes based on the exception type.
- **Overflow Handling:** The code doesn't explicitly handle integer overflow. It might be beneficial to add checks and potentially set an overflow flag.

+++ src/main/kotlin/environment/errorsCatchable/ErrorType.kt
@@ -1,6 +1,6 @@
 package environment.errorsCatchable
 
-enum class ErrorType(val code: Byte) {
+enum class ErrorType(val code: Long) {
 	FILE_NOT_FOUND(0),
 	FILE_ALREADY_EXISTS(1),
 	LONG_OVERFLOW(2),
